ubah beberapa
