+++ PosTekpro/src/postekpro/Isi.java
@@ -2,6 +2,9 @@ package postekpro;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
+import java.util.ArrayList;
+import java.util.List;
+
 
 public class Isi {
 
@@ -13,7 +16,7 @@ public class Isi {
     private long saldo;
 
     public String getNama() {
-        return nama;
+        return this.nama;
     }
 
     public void setNama(String nama) {
@@ -21,11 +24,11 @@ public class Isi {
     }
 
     public long getSaldo() {
-        return saldo;
+        return this.saldo;
     }
 
     public void setSaldo(Integer saldo) {
         this.saldo = saldo;
     }
-
+    
 }
